requirements.txt
pyspark==3.5.4
scikit-learn==1.6.0
mlflow==2.19.0
fastapi==0.115.6
seaborn==0.13.2
minio==7.2.13
findspark==2.0.1
imbalanced-learn==0.13.0
torch==2.5.1
skl2onnx==1.18.0
onnx==1.17.0
onnxruntime==1.20.1
psycopg2==2.9.10
loguru==0.7.3
python-dotenv==1.0.1
confluent-kafka==2.8.0